% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comid_to_huc}
\alias{comid_to_huc}
\alias{huc_to_comids}
\title{Convert COMID to HUC and vice versa}
\usage{
comid_to_huc(comids, huc.level = 6)

huc_to_comids(hucs)
}
\arguments{
\item{comids}{Any valid comid. See huc_comid_xref$comid for options}

\item{huc.level}{The huc.level to return. Defaults to 6 (i.e. HUC6)}
}
\value{
A vector of resulting HUCs (for \code{comid_to_huc}) or COMIDs
(for \code{huc_to_comid})
}
\description{
These are simple helper functions that take either a comid or huc as an
argument and return the other. COMIDs are nested with HUCs, so
\code{comid_to_huc} is 1-to-1, but \code{huc_to_comid} is 1-to-many.
}
\examples{
comid_to_huc(23940175)
}
