% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_data.R
\name{download_data}
\alias{download_data}
\alias{download_huc_data}
\alias{download_all_hucs}
\title{Download stream temperature data for a specific HUC-6 watershed}
\usage{
download_huc_data(hucs = NULL, comids = NULL, timeout = 300)

download_all_hucs(timeout = 1000)
}
\arguments{
\item{hucs}{A HUC watershed. Can be any level--will get truncated to HUC6}

\item{comids}{Any COMID that can be found in \code{huc_comid_xref}}

\item{timeout}{If timeout errors are experienced on download, this argument
can be increased}
}
\value{
Downloads a file from (Siegel et al. 2023. Daily stream temperature
predictions for free-flowing streams in the Pacific
Northwest)\link{https://zenodo.org/records/8174951}
}
\description{
This function will download data from
\href{https://zenodo.org/records/8174951}{Siegel et al. 2023. Daily stream temperature predictions for free-flowing streams in the Pacific Northwest},
extract it, and embed it within the package directory structure so that it
is accessible by \code{\link{read_stream_temps}}
}
\examples{
\dontrun{
  # these download files of .75 to 4GB, so can take a bit
  download_huc_data(hucs = 170103)
  download_huc_data(comid = 23940175)
}
}
