% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_stream_temps.R
\name{read_stream_temps}
\alias{read_stream_temps}
\title{Read stream temperature data for a specific COMID or HUC and date range}
\usage{
read_stream_temps(comid = NULL, start.date = NULL, end.date = NULL, huc = NULL)
}
\arguments{
\item{comid}{The NDHPlus COMID for a stream reach. Anything found in
huc_comid_xref$comid is a valid option}

\item{start.date}{Start date entered as "YYYY-MM-DD". If left NULL (default)
it will either return everything prior to the end.date or the full range of
dates (if end.date = NULL)}

\item{end.date}{End date entered as "YYYY-MM-DD". If left NULL (default)
it will either return everything after the start.date or the full range of
dates (if start.date = NULL)}

\item{huc}{Optionally. Can enter any HUC from HUC6 to HUC10. Anything found
in huc_comid_xref$huc is a valid option. Entering a huc without a comid
will return all COMIDs for that HUC level (be careful...you might get a lot
of data back)}
}
\value{
A data.frame of predicted stream temperatures by date for each
comid
}
\description{
Read stream temperature data for a specific COMID or HUC and date range
}
\examples{
\dontrun{
 temps <- read_stream_temps(comid = 22878865, start.date = "2015-01-01", end.date = "2015-12-31")
 temps2 <- read_stream_temps(comid = 22878865, start.date = "2019-01-01")
 temps3 <- read_stream_temps(comid = 22878865, end.date = "1990-03-31")
}
}
